Choose photo from library or take from camera, and then send to LLM backend

+++ apps/ios/VoiceAssistant/types/common.swift
@@ -21,8 +21,8 @@ enum ImageFormat: String, Codable {
 }
 
 struct Image: Codable {
-    let width: Int
-    let height: Int
+    let width: CGFloat
+    let height: CGFloat
     let format: ImageFormat
     let data: Data
